refactor Order,payment 도메인 리팩토링 진행 #13

+++ src/main/java/kr/hhplus/be/server/application/order/OrderItemCommand.java
@@ -3,5 +3,5 @@ package kr.hhplus.be.server.application.order;
 //상품 단위 요청
 public record OrderItemCommand (
         Long productId,
-        int quantity
+        Long quantity
 ){}
